# spec/openapi-dashboard.yml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: E2B Dashboard API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminJWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    projectID:
      name: projectID
      in: path
      required: true
      description: Identifier of the project.
      schema:
        type: string
        format: uuid
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "501":
      description: Operation is not implemented
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code.
        message:
          type: string
          description: Error message.
...
    ManagementProjectLimits:
      type: object
      description: >-
        A project's effective limits, already resolved by the caller. Every
        field is absolute: this side stores what it is given and performs no
        arithmetic of its own.


        The minimums below track the CHECK constraints on tiers, which is the
        contract for what a limit may be. project_limits stores the same values
        under looser constraints on purpose — it is a push target, and a floor
        that only rejects the impossible keeps a future decision about what is
        allowed a change to this schema rather than a migration.


        `max_disk_size_mb` and `max_free_disk_size_mb` are one ceiling under two
        names, and either name alone carries it. A caller that sends both must
        send them equal; a caller that sends neither is refused. Sending both is
        what a caller does while receivers older than the second name are still
        running.
      required:
        - revision
        - concurrent_sandboxes
        - max_sandbox_length_hours
        - max_vcpu
        - max_ram_mb
        - disk_mb
        - concurrent_template_builds
        - events_ttl_days
        - default_free_disk_size_mb
        - api_team_rps_list
      anyOf:
        - required: [max_free_disk_size_mb]
        - required: [max_disk_size_mb]
      properties:
        revision:
          type: integer
          format: int64
          minimum: 1
          description: >-
            The caller's version of this answer, raised whenever the limits it
            resolved for the project change. Delivery is over a network, so two
            pushes can be in flight at once and arrive in either order: this
            side stores the revision it accepted and drops a delivery at or
            below it, which is what keeps a delayed retry from putting the
            project back on limits it has already left.


            Comparable only against earlier revisions for the same project.
        concurrent_sandboxes:
          type: integer
          format: int32
          minimum: 1
        max_sandbox_length_hours:
          type: integer
          format: int32
          minimum: 1
        max_vcpu:
          type: integer
          format: int32
          minimum: 1
        max_ram_mb:
          type: integer
          format: int64
          minimum: 1
        disk_mb:
          type: integer
          format: int64
          minimum: 1
        concurrent_template_builds:
          type: integer
          format: int32
          minimum: 1
        events_ttl_days:
          type: integer
          format: int32
          minimum: 1
        default_free_disk_size_mb:
          type: integer
          format: int64
          minimum: 0
          description: >-
            The default free-space growth target when a template build request
            omits one. May sit anywhere at or below the maximum free-space
            growth target, and usually sits well below it; a delivery whose
            default exceeds the maximum is rejected.
        max_free_disk_size_mb:
          type: integer
          format: int64
          minimum: 1
          description: >-
            The most a template build may request as its free-space growth
            target.
        max_disk_size_mb:
          type: integer
          format: int64
          minimum: 1
          description: >-
            The same ceiling as max_free_disk_size_mb, under the name it was
            first published with. Kept until every deployed sender and receiver
            speaks the other name.
        api_team_rps_list:
          type: integer
          format: int64
          minimum: 0
          description: >-
            Requests per second a project may make to each list endpoint. Zero
            disables the limit.
...
paths:
...
  /v1/management/projects/{projectID}/limits:
    parameters:
      - $ref: "#/components/parameters/projectID"
    put:
      operationId: managementUpsertProjectLimits
      summary: Reconcile effective limits for a project (v1).
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementProjectLimits"
      responses:
        "204":
          description: Effective limits are synchronized.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "501":
          $ref: "#/components/responses/501"
...